+++ internal/cloudflare.go
@@ -22,7 +22,7 @@ func SetupCloudflareClient() {
 // GetCloudflareZoneID creates a global var containing the zone id
 func GetCloudflareZoneID() {
 	var err error
-	zoneID, err = cf.ZoneIDByName(config.RootDomain)
+	zoneID, err = cf.ZoneIDByName(config.Cloudflare.ZoneName)
 	if err != nil {
 		log.Fatalf("Unable to get Cloudflare zone id: %s", err)
 	}
@@ -30,16 +30,21 @@ func GetCloudflareZoneID() {
 }
 
 // GetCloudflareDNSRecords gathers all DNS records in a given zone
-func GetCloudflareDNSRecords() []cloudflare.DNSRecord {
+func GetCloudflareDNSRecords() ([]cloudflare.DNSRecord, []cloudflare.DNSRecord) {
 	dnsRecords, err := cf.DNSRecords(zoneID, cloudflare.DNSRecord{})
 	if err != nil {
 		log.Fatalf("Unable to get Cloudflare DNS records: %s", err)
 	}
 
 	var cloudflareDNSRecords []cloudflare.DNSRecord
+	var deleteGraceRecords []cloudflare.DNSRecord
 	var cloudflareDNSRecordNames []string
 
 	for _, dnsRecord := range dnsRecords {
+		if dnsRecord.Type == "TXT" && strings.Contains(dnsRecord.Name, "_syncflaer._deletegrace") {
+			deleteGraceRecords = append(deleteGraceRecords, dnsRecord)
+			continue
+		}
 		if dnsRecord.Type != "CNAME" && dnsRecord.Type != "A" {
 			continue
 		}
@@ -47,68 +52,84 @@ func GetCloudflareDNSRecords() []cloudflare.DNSRecord {
 		cloudflareDNSRecordNames = append(cloudflareDNSRecordNames, dnsRecord.Name)
 	}
 	log.Debugf("Found Cloudflare DNS records: %s", strings.Join(cloudflareDNSRecordNames, ", "))
-	return cloudflareDNSRecords
+	return cloudflareDNSRecords, deleteGraceRecords
 }
 
 // CreateCloudflareDNSRecord is a wrapper function to create a DNS record
 func CreateCloudflareDNSRecord(record cloudflare.DNSRecord) {
-	newDNSRecord := cloudflare.DNSRecord{
-		Type:    record.Type,
-		Name:    record.Name,
-		Content: record.Content,
-		Proxied: record.Proxied,
-		TTL:     record.TTL,
-	}
-
-	_, err := cf.CreateDNSRecord(zoneID, newDNSRecord)
+	_, err := cf.CreateDNSRecord(zoneID, record)
 	if err != nil {
-		addSlackMessage(fmt.Sprintf("Unable to create DNS record %s: %s", newDNSRecord.Name, err), "danger")
-		log.Errorf("Unable to create DNS record %s: %s", newDNSRecord.Name, err)
+		var errMsg string
+		if record.Type == "A" || record.Type == "CNAME" {
+			errMsg = fmt.Sprintf("Unable to create DNS record %s: %s", record.Name, err)
+		}
+		if record.Type == "TXT" {
+			errMsg = fmt.Sprintf("Unable to create delete grace DNS record %s: %s", record.Name, err)
+		}
+		addSlackMessage(errMsg, "danger")
+		log.Error(errMsg)
 		return
 	}
 
-	addSlackMessage(fmt.Sprintf("Created: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", newDNSRecord.Name, newDNSRecord.Type, newDNSRecord.Content, newDNSRecord.Proxied, newDNSRecord.TTL), "good")
-	log.Infof("Created: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", newDNSRecord.Name, newDNSRecord.Type, newDNSRecord.Content, newDNSRecord.Proxied, newDNSRecord.TTL)
+	infoMsg := fmt.Sprintf("Created: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", record.Name, record.Type, record.Content, record.Proxied, record.TTL)
+	if record.Type == "A" || record.Type == "CNAME" {
+		addSlackMessage(infoMsg, "good")
+		log.Info(infoMsg)
+	}
+	if record.Type == "TXT" {
+		log.Debug(infoMsg)
+	}
 }
 
 // DeleteCloudflareDNSRecord is a wrapper function to delete a DNS record
 func DeleteCloudflareDNSRecord(record cloudflare.DNSRecord) {
 	err := cf.DeleteDNSRecord(zoneID, record.ID)
 	if err != nil {
-		addSlackMessage(fmt.Sprintf("Unable to delete DNS record %s: %s", record.Name, err), "danger")
-		log.Errorf("Unable to delete DNS record %s: %s", record.Name, err)
+		var errMsg string
+		if record.Type == "A" || record.Type == "CNAME" {
+			errMsg = fmt.Sprintf("Unable to delete DNS record %s: %s", record.Name, err)
+		}
+		if record.Type == "TXT" {
+			errMsg = fmt.Sprintf("Unable to delete delete grace DNS record %s: %s", record.Name, err)
+		}
+		addSlackMessage(errMsg, "danger")
+		log.Error(errMsg)
 		return
 	}
 
-	addSlackMessage(fmt.Sprintf("Deleted: %s", record.Name), "good")
-	log.Infof("Deleted: %s", record.Name)
+	infoMsg := fmt.Sprintf("Deleted: %s", record.Name)
+	if record.Type == "A" || record.Type == "CNAME" {
+		addSlackMessage(infoMsg, "good")
+		log.Info(infoMsg)
+	}
+	if record.Type == "TXT" {
+		log.Debug(infoMsg)
+	}
 }
 
-// UpdateCloudflareDNSRecords updates the public IP and additionalRecords
-func UpdateCloudflareDNSRecords(cloudflareDNSRecords []cloudflare.DNSRecord, userRecords []cloudflare.DNSRecord) {
-	for _, dnsRecord := range cloudflareDNSRecords {
-		for _, userRecord := range userRecords {
-			if dnsRecord.Name != userRecord.Name {
-				continue
-			}
-			if dnsRecord.Proxied == userRecord.Proxied && dnsRecord.TTL == userRecord.TTL && dnsRecord.Content == userRecord.Content {
-				continue
-			}
-			updatedDNSRecord := cloudflare.DNSRecord{
-				Type:    userRecord.Type,
-				Content: userRecord.Content,
-				Proxied: userRecord.Proxied,
-				TTL:     userRecord.TTL,
-			}
-			err := cf.UpdateDNSRecord(zoneID, dnsRecord.ID, updatedDNSRecord)
-			if err != nil {
-				addSlackMessage(fmt.Sprintf("Unable to update DNS record %s: %s", dnsRecord.Name, err), "danger")
-				log.Errorf("Unable to update DNS record %s: %s", dnsRecord.Name, err)
-				continue
-			}
-			addSlackMessage(fmt.Sprintf("Updated: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", dnsRecord.Name, updatedDNSRecord.Type, updatedDNSRecord.Content, updatedDNSRecord.Proxied, updatedDNSRecord.TTL), "good")
-			log.Infof("Updated: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", dnsRecord.Name, updatedDNSRecord.Type, updatedDNSRecord.Content, updatedDNSRecord.Proxied, updatedDNSRecord.TTL)
+// UpdateCloudflareDNSRecord is a wrapper function to update a DNS record
+func UpdateCloudflareDNSRecord(record cloudflare.DNSRecord) {
+	err := cf.UpdateDNSRecord(zoneID, record.ID, record)
+	if err != nil {
+		var errMsg string
+		if record.Type == "A" || record.Type == "CNAME" {
+			errMsg = fmt.Sprintf("Unable to update DNS record %s: %s", record.Name, err)
 		}
+		if record.Type == "TXT" {
+			errMsg = fmt.Sprintf("Unable to update delete grace DNS record %s: %s", record.Name, err)
+		}
+		addSlackMessage(errMsg, "danger")
+		log.Error(errMsg)
+		return
+	}
+
+	infoMsg := fmt.Sprintf("Updated: name: %s, type: %s, content: %s, proxied: %t, ttl: %d", record.Name, record.Type, record.Content, record.Proxied, record.TTL)
+	if record.Type == "A" || record.Type == "CNAME" {
+		addSlackMessage(infoMsg, "good")
+		log.Info(infoMsg)
+	}
+	if record.Type == "TXT" {
+		log.Debug(infoMsg)
 	}
 }
 
@@ -146,4 +167,25 @@ func GetOrphanedDNSRecords(cloudflareDNSRecords []cloudflare.DNSRecord, userReco
 		}
 	}
 	return orphanedRecords
+}
+
+// UpdateOutdatedDNSRecords checks whether the DNS records must be updated
+func UpdateOutdatedDNSRecords(cloudflareDNSRecords []cloudflare.DNSRecord, userRecords []cloudflare.DNSRecord) {
+	for _, dnsRecord := range cloudflareDNSRecords {
+		for _, userRecord := range userRecords {
+			if dnsRecord.Name != userRecord.Name {
+				continue
+			}
+			if dnsRecord.Proxied == userRecord.Proxied && dnsRecord.TTL == userRecord.TTL && dnsRecord.Content == userRecord.Content {
+				continue
+			}
+			updatedDNSRecord := cloudflare.DNSRecord{
+				Type:    userRecord.Type,
+				Content: userRecord.Content,
+				Proxied: userRecord.Proxied,
+				TTL:     userRecord.TTL,
+			}
+			UpdateCloudflareDNSRecord(updatedDNSRecord)
+		}
+	}
 }